Force MPSC channels for API.
