+++ webrender_api/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2018"
 
 [features]
 nightly = ["euclid/unstable", "serde/unstable"]
-ipc = ["ipc-channel"]
 serialize = []
 deserialize = []
 display_list_stats = []
@@ -19,7 +18,6 @@ app_units = "0.7"
 bitflags = "1.2"
 byteorder = "1.2.1"
 derive_more = "0.13"
-ipc-channel = {version = "0.12.0", optional = true}
 euclid = { version = "0.20.0", features = ["serde"] }
 malloc_size_of_derive = "0.1"
 serde = { version = "1.0", features = ["rc"] }
